# Conflicts:
#	pom.xml
#	src/main/java/net/sf/marineapi/ais/parser/AISMessageParser.java
#	src/main/java/net/sf/marineapi/nmea/parser/SentenceFactory.java
#	src/main/java/net/sf/marineapi/nmea/sentence/SentenceId.java

+++ src/main/java/net/sf/marineapi/ais/message/AISMessage05.java
@@ -115,7 +115,7 @@ public interface AISMessage05 extends AISMessage {
 	 * Returns the maximum draught.
 	 * @return an integer value of the maximum static draught in 1/10 m
 	 */
-	int getMaximumDraught();
+	double getMaximumDraught();
 
 	/**
 	 * Returns the destination.
